requirements-frozen.txt
alembic==1.4.2
amqp==2.6.1
aniso8601==8.0.0
apispec==3.3.2
apispec-webframeworks==0.5.2
appdirs==1.4.4
attrs==20.1.0
-e git+git@github.com:baotran2207/python_design.git@7cf8171bbb9daa5f9f27dc445392928e389958f8#egg=baoapi
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
distlib==0.3.1
factory-boy==3.0.1
Faker==4.1.2
filelock==3.0.12
Flask==1.1.2
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
humanize==2.6.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.5.0
packaging==20.4
passlib==1.7.2
pgcli==3.0.0
pgspecial==1.11.10
pluggy==0.13.1
prompt-toolkit==3.0.7
psycopg2==2.8.5
py==1.9.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-factoryboy==2.0.3
pytest-flask==1.0.0
pytest-runner==5.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
tabulate==0.8.7
terminaltables==3.1.0
text-unidecode==1.3
toml==0.10.1
tox==3.20.0
twilio==6.45.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
